+++ src/commands/Management/DeleteGameFromCRSchedule.js
@@ -1,18 +1,11 @@
 import {
-    Command
+  Command
 } from "@sapphire/framework";
 import {
-    container
+  container
 } from '@sapphire/pieces';
-import {
-    bold,
-    green,
-    logthis,
-    red
-} from "../../lib/hercConfig.js";
-import {
-    AddToUsageStats
-} from '../../utils/SQL/AddToUsageStats.js';
+import { statcord } from "../../Hercules.js";
+import { LogGreen, LogRed } from "../../utils/ConsoleLogging.js";
 const OBJCrgameSched = container.OBJCrgameSched
 const GameScheduleManager = container.hercGameSchedMngr;
 export class DeleteScheduledGame extends Command {
@@ -27,20 +20,20 @@ export class DeleteScheduledGame extends Command {
   }
 
   async messageRun(message, args) {
-    //* -------------- */
+
     //* Adding to usage stats
     var userid = message.author.id
-    var SQLTargetTable = `deletegamefromcrschedulestats`
-    var commandname = `deletecr`
-    AddToUsageStats(userid, SQLTargetTable, commandname)
+    var commandname = `deletecreatedgame`
+    statcord.postCommand(commandname, userid)
     //* -------------- */
+
     const text = await args.rest('string').catch(() => null);
     var selectedScheduledGame = OBJCrgameSched[`${text}`]
-    logthis(red(bold(`[Game Scheduling] Removing: ${selectedScheduledGame} from the Game Schedule`)))
+    LogRed(`[Game Scheduling] Removing: ${selectedScheduledGame} from the Game Schedule`)
     try {
       message.reply(`Deleted: '${selectedScheduledGame}' from today's channel scheduling.`)
       GameScheduleManager.deleteJob(selectedScheduledGame)
-      logthis(green(bold(`[Game Scheduling] Successfully removed ${selectedScheduledGame} from the Game Schedule.`)))
+      LogGreen(`[Game Scheduling] Successfully removed ${selectedScheduledGame} from the Game Schedule.`)
     } catch (error) {
       console.log(error)
       return