+++ src/commands/Utility/standings.js
@@ -2,9 +2,9 @@ import {
   Command
 } from '@sapphire/framework';
 import {
-  AddToUsageStats
-} from '../../utils/SQL/AddToUsageStats.js';
-export class PostStandings extends Command {
+  statcord
+} from '../../Hercules.js';
+export class Standings extends Command {
   constructor(context, options) {
     super(context, {
       ...options,
@@ -14,25 +14,27 @@ export class PostStandings extends Command {
   }
 
   async messageRun(message, args) {
-    //* -------------- */
     //* Adding to usage stats
     var userid = message.author.id
-    var SQLTargetTable = `standingsstats`
     var commandname = `standings`
-    AddToUsageStats(userid, SQLTargetTable, commandname)
+    statcord.postCommand(commandname, userid)
     //* -------------- */
+
     const msg = message;
     //? catching no arguments - which we will use here to post BOTH standings
     const text = await args.rest('string').catch(() => null);
+
     if (text == null) {
       msg.reply("Here are the Eastern and Western NBA Conference Standings:")
       msg.reply({
-        files: ['./gameimages/eaststandings.jpg',
+        files: [
+          './gameimages/eaststandings.jpg',
           './gameimages/weststandings.jpg'
         ]
       });
       return;
     }
+
     if (text.toLowerCase() == "west") {
       msg.reply("Here are the Western Conference Standings")
       msg.reply({