+++ src/commands/VerifyServer.js
@@ -4,8 +4,9 @@ import {
 import {
   container
 } from '@sapphire/pieces';
-import { bold, green, logthis, noDbLoadedMsg, yellow } from '../lib/hercConfig.js';
-import { SendEmbedErrorResp } from '../utils/SQL/Embeds/ErrorReplyEmbed.js';
+import { noDbLoadedMsg } from '../lib/hercConfig.js';
+import { LogGreen } from '../utils/ConsoleLogging.js';
+import { SendEmbedErrorResp } from '../utils/Send Embeds/ErrorReplyEmbed.js';
 import { VerifyDatabase } from '../utils/VerifyDatabase.js';
 export class verifyserver extends Command {
   constructor(context, options) {  
@@ -21,13 +22,12 @@ export class verifyserver extends Command {
 
   async messageRun(message){  
     var serverlocalornot = container.WhichServer;
-    logthis(yellow(bold(`[DEBUGGING] Verifying Server - The ${serverlocalornot} database is currently loaded.`)))
     if (VerifyDatabase() == 'No Database is loaded') {
-      logthis(green(bold(serverlocalornot)))
+      LogGreen(`[verifyserver.js] No database is currently loaded.`)
         message.reply({embeds: [SendEmbedErrorResp(noDbLoadedMsg)]})
       return;
     } else {
-      logthis(green(bold(serverlocalornot)))
+      LogGreen(`[verifyserver.js] Database ${serverlocalornot} is loaded.`)
     message.reply(`The ${serverlocalornot} database is currently loaded.`)
     }
   }